Integrate Alpenhorn into the Vuvuzela client
The vuvuzela-client command now requires the -username flag, which
expects a username that has been initialized with vuvuzela-keygen. The
-pki flag is still required to provide server information for the convo
service, but will be removed in a future commit after we overhaul the
convo mixnet.

This replaces the /talk command with two new commands: /addfriend and
/call. The /addfriend command sends a friend request to another user.
The /call command initiates a conversation with an existing friend.
For now, the client accepts all friend requests and calls.