all: rewrite mixnet and entry (coordinator) server
The new mixnet package implements the Vuvuzela conversation protocol and
is based on the Alpenhorn mixnet design. Now mixers are identified by
their long-term signing keys and the onion keys are forward secure.

The coordinator package replaces the Vuvuzela entry server and is based
on the Alpenhorn coordinator. The Vuvuzela client is updated to speak
with the new coordinator.

The vuvuzela-mixer command replaces vuvuzela-server and the
vuvuzela-coordinator command replaces vuvuzela-entry-server.

Alpenhorn's config package is now used for global server configuration,
so the PKI type is removed. This also deletes code that was only used by
the previous server implementations.