coordinator: dynamically configure clients
The coordinator now broadcasts a NewRound message at the start of every
round to let clients know the configuration for that round. Configs are
managed by an underlying config.Server that is loaded from the
ConfigServerPersistPath. The coordinator still has its own persistent
state to keep track of the last round but its format has changed. The
init process is also revamped to match the Alpenhorn coordinator -- both
coordinators can be bootstrapped from the same bootstrap config file.
In the future, we may merge the two coordinators to reduce code
duplication.

This commit brings the coordinator in sync with the previous mixnet
commit. Also, the HTTP routes have changed and clients must now connect
to the coordinator over HTTPS (edTLS).

This commit is based on commits to Alpenhorn:

44d2d99fcd3b8eb73e2d0dfb946cc66aef5f8a4a  coordinator: dynamically configure clients
7a0913486971623216ed5946bf171fc1326ac459  coordinator: use the new config server

Those commits give a more detailed explanation of these changes and also
talk about issues related to freshness of the configs.

The client will be updated in the next commit.