client: handle config updates; new client package
This commit syncs the client with recent changes to the coordinator and
mixnet. The client now handles config updates sent by the coordinator
via the newround message. Config updates need to be persisted for
security, so now the client has persistent state.

Previously, the Client type was part of the vuvuvzela-client package.
Now, the vuvuzela-keygen command also needs access to the Client type
so that it can bootstrap the client with an initial convo config. This
commit moves the Client to its own package so that both commands can
use it.