coordinator: allow multiple onions from clients
This is the first step in enabling the Vuvuzela client to support
multiple conversations. Now clients can send multiple onions in a
single "onion" message, and the coordinator's "reply" message will
contain the corresponding response onions.

The typesocket maxMessageSize will likely need to be increased to
support larger messages.

The client and GUI will be updated in followup commits.