client: support multiple conversations
Previously, a ConvoHandler represented a single conversation that could
be switched in and out of the client. Now there is just one ConvoHandler
provided to the client when it is created. The new ConvoHandler supports
multiple messages per round, but it is up to the handler to determine
how these map to conversations. The ConvoHandler should create the same
number of dead drop messages each round, but the client does not enforce
this. In that sense, the client is a low-level interface to the convo
protocol.

The GUI client will be updated in a followup commit.