cmd/vuvuzela-client: support multiple conversations
The GUI client now supports multiple concurrent conversations. There can
be at most 5 active conversations. The client will generate cover
traffic if there are fewer than 5 active conversations.

The /call command is used to activate a conversation and invokes
Alpenhorn's dialing protocol. The /hangup command ends a conversation
and frees up slot to activate other conversations. If the client
receives a call but is at the limit of active convos, then the client
will hold the call. The user can answer a held call using the /answer
command.

Each conversation now has its own window (or view as gocui calls it).
Users can create (or jump to) a conversation window with the /w command
and close a window with /wc. Like in irssi, each window has a number and
users can move between windows with Alt-{1,2,3,...}. The status bar now
features a menu bar that signals the currently focused window and any
windows with unread messages. The first window (window 1) is the main
window and is used for non-convo-specific messages like friend requests.
In code, this window is a special case since it does not map to a
Conversation value.

The concurrency in the GUI client is a bit of a mess. For example, there
are now two ways to write to a window, Printf and PrintfSync, depending
on which Go routine the call is made from. Also, having a mutex per
conversation adds complexity and has been a source of deadlocks. The
deadlock package has been useful in debugging deadlocks:

    https://github.com/sasha-s/go-deadlock