cmd/vuvuzela-client: implement forward secrecy for convo messages
The idea is to use a keywheel (like in Alpenhorn's dialing protocol) to
achieve forward secrecy for convo messages. Previously, the dialing
protocol was used to agree on a session key that was fixed for the
duration of the conversation. Now the session key is rotated every convo
round via a keywheel. However, to bootstrap the keywheel, users need to
agree on the convo round that the initial session key corresponds to.

One naive solution is to use the latest convo round rounded down to the
nearest multiple of 1000. This works most of the time, except when a
call is sent near a rounding boundary (then the sender might think the
keywheel starts at round x, while the receiver thinks it starts at
x+1000).

Our solution is to use dialing intents to indicate which side of the
rounding boundary the sender intends to use. We use two intents: 0 means
that the rounded down convo round is even, and 1 means that it is odd.
Assuming the sender's and receiver's convo rounds are loosely synced,
then this is sufficient for them to agree on a convo round. The new
roundSyncer type implements this logic.

Fixes #3.