cmd/vuvuzela-client: improve latency calculation
Previously, the latency displayed in the GUI was accurate only if your
clock was in sync with your chat partner's clock, but this is rarely
the case. Now the latency is computed using only your own clock.

We leave TimestampMessage in the code as a reference for adding new
types of messages later.