cmd/vuvuzela-client: overhaul the getting started process
This overhauls how the user gets started using Vuvuzela, including key
generation and registering with the PKGs.

This removes the vuvuzela-keygen command. Instead, the Vuvuzela client
generates the long-term key when it is launched for the first time.

Users should now use the /register command to register with PKGs after
obtaining a registration token (e.g., from vuvuzela.io). The /connect
command is only used to retry connecting if it fails at startup.