coordinator,client: make coordinator loop more efficient
The coordinator now starts multiple rounds in parallel and relies on the
round's end time to schedule the rounds. This allows the coordinator to
prep the mixers and clients for a round well in advance of clients
submitting their onions. We also avoid fetching the latest config every
round, and instead fetch it every minute. This change also combines the
MixRound and NewRound messages to avoid an unnecessary round-trip. As a
result, clients will need to be updated before reconnecting to the Convo
service.

Overall, this commit significantly improves the end-to-end latency of
the convo protocol. However, there are several new knobs in this design
that will be tweaked as we get experience with it in production.