cmd/vuvuzela-client: implement message retransmission
The client cycles through the out queue to avoid unnecessary message
retransmission. The client also prints incoming messages in sequence
number order, queueing messages that come in out of order. The sequence
number 0 is reserved for cover traffic.

The client uses the first non-zero sequence number it sees as the
starting ack value, but this means that messages can be lost at the
start of a conversation.

Updates issue #5.