cmd/vuvuzela-client: be more precise in adjusting lastOut on ack
Previously, the sending sequence would restart from the lowest message
when anything got ACKed.  In steady-state when there's no message
loss, this resulted in stuttering: each message would get sent twice.
When message 1 got acked, message 2 was already in-flight (and would have
been acked soon), but resetting lastOut caused message 2 to be sent again.

The new behavior is to continue in the looping order of outstanding
messages.  In steady-state with no message loss, this should result in
2x higher throughput by avoiding stuttering.